setup fabric entry point with yumi mc foundation
