+++ fabric/build.gradle
@@ -5,9 +5,19 @@ plugins {
 }
 
 dependencies {
-    minecraft "com.mojang:minecraft:${minecraft_version}"
-    implementation "net.fabricmc:fabric-loader:${fabric_loader_version}"
-    implementation "net.fabricmc.fabric-api:fabric-api:${fabric_version}"
+    minecraft "com.mojang:minecraft:${project.minecraft_version}"
+    implementation "net.fabricmc:fabric-loader:${project.fabric_loader_version}"
+    implementation "net.fabricmc.fabric-api:fabric-api:${project.fabric_version}"
+
+    implementation "org.ladysnake.cardinal-components-api:cardinal-components-base:${project.cca_version}"
+    implementation "org.ladysnake.cardinal-components-api:cardinal-components-entity:${project.cca_version}"
+
+    compileOnly "eu.pb4:polymer-core:${project.polymer_version}"
+    compileOnly "eu.pb4:polymer-networking:${project.polymer_version}"
+
+    testImplementation "net.fabricmc:fabric-loader-junit:${project.fabric_loader_version}"
+    testImplementation "org.junit.jupiter:junit-jupiter"
+    testImplementation "org.mockito:mockito-core:3.+"
 }
 
 loom {
