+++ gradle.properties
@@ -29,3 +29,4 @@ neoforge_loader_version_range=[4,)
 cca_version=8.0.0
 polymer_version=0.16.5+26.1.2
 yumi_version=2.0.0
+yumi_mc_foundation_version=1.1.0+26.1
